pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pothole_report"]

[project]
name = "pothole-report"
version = "0.1.0"
description = "Extract GPS from pothole photos and output postcode, map link, and description for UK Fill That Hole reporting"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "geopy>=2.4.1",
    "keyring>=25.7.0",
    "pillow>=12.1.0",
    "pyyaml>=6.0.3",
    "rich>=14.3.1",
]

[project.scripts]
report-pothole = "pothole_report.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]
